🔥 chore: move src/theme/colors.ts to src/config/colors.ts
