+++ src/theme/index.ts
@@ -1,5 +1,12 @@
 import { createTheme } from "@mui/material/styles";
-import { BLACK, BLACK_OFF, RED, RED_LIGHT, TAN, WHITE } from "src/theme/colors";
+import {
+  BLACK,
+  BLACK_OFF,
+  RED,
+  RED_LIGHT,
+  TAN,
+  WHITE,
+} from "src/config/colors";
 import ButtonTheme from "./Button";
 
 const theme = createTheme({
